requirements.txt
aiogram==3.0.0b7
asyncio
python-dotenv
logging
motor==2.5.1
pandas==2.1.1
openai
pymongo==3.12.0
requests==2.31.0
telethon==1.24.0
